Revert "app: use FAST filter when painting xfer surface"
This reverts commit 71566e7ba4627509fffb9535d423ca2646fb4364 [formerly 36258a671a04336260b1154128f8ff2cd9b85673].
This commit was making the rotated canvas rendering quite horrible to
the point that I think it would make the canvas rotation feature barely
usable. See bug 759287, comments 13 to 18.
I think we will need to find other ways to accelerate rendering.
Compromise on quality is possible, but I think that in this case, this
was more than just a compromise. It was more like completely abandonning
quality. We could even see the lines "spiking" while you were rotating!
Like your drawing was alive!


Former-commit-id: 21cb905dad8e8ccb95a4a420293102a478b7638e