app: run gimp_projectable_get_graph() before gegl_node_get_parent()...
... on top-level layers.
There was even a comment for this, but I missed this when I moved some
code to the top of the function in commit 5cd193762b [formerly b9577a783d]. Now moving this
call up as well. This appeared to be more of a problem when merging
layers without a GUI (script-fu). I'm guessing the GUI calls this anyway
before.


Former-commit-id: b4c7dd8fd809b817a6e24c7f8d8a57003017146d