app: use gimp_transform_polygon() in gimp_transform_resize_boundary()
... so that the transformed boundary is properly clipped.

Adjust the boundary-size algorithms to operate on arbitrary
polygons.

Avoid using gimp_matrix3_will_explode() in
gimp_drawable_transform_buffer_affine() and falling back to
cropping the result, and avoid setting the "clip-to-input" property
of gegl:transform.  Neither of those in needed anymore.

This effectively reverts the app/ part of commit
206fa93c4d7265b944b11c9237703e07783557f7 [formerly 768d06614f203bf555bbda1f6186d4730ae2f8b5].  The next commit revets
the libgimpmath/ part.


Former-commit-id: 3c0787e4c360348850df1898cca25c5db9ac967f