Bug 779839 - Add <releases>-Tag to the gimp.appdata.xml.in.
- appstream-util returns a "style-invalid" error: "<ul> cannot start a
  description [(null)]". So I add a <p> introduction to the 2.9.8
  <release> tag. This was part of unit test failure on the appdata file.
- I also add a type property for 2.9.8. This is a new property which I
  proposed and which just got accepted in the appstream specification:
  https://github.com/ximion/appstream/pull/158
- I add <release> tags for all previous 2.9.x releases. No description
  for these, just a type property. But feel free to propose patches
  adding short non-technical description for these.

Note: it was originally proposed in the bug report to use the appdata
file in place of NEWS (and have this one generated from appdata). But
after discussion with appstream project, appdata is expected to be
concise, non-technical and more "marketing" than exhaustive. This is
quite a different usage than NEWS which is more an exhaustive summary of
new features and major changes. So these 2 files will likely remain
distinct.

Former-commit-id: 78f251b0a745219b2f119dc093b2744049d25f17