app: when WARNING or CRITICAL debugging are ignored, go to terminal.
Current code was redirecting WARNING and CRITICAL errors to normal
messaging when the debugging was deactivated (in Preferences). But if
you deactivate these on purpose, then it means you don't want to get
annoyed by small pop-ups either.
This commit makes them directly displayed in terminal, as they used to
before, when debugging is deactivated.


Former-commit-id: 37f99069616d3d20939808dbc6eb5af6da505e7e