Bug 784802 - Crop and rectangle-select tools incorrectly detect ...
... current aspect ratio

When starting the rectangle-select (and ellipse-select) tools,
properly reset their default aspect ratio to 1:1.  This fixes an
issue where the tool would use the aspect ratio of the last
rectangle when there's no user-overriden aspect ratio specified.

This restores the 2.8 behavior, except for the fact that the aspect
ratio resets to 1:1 when the tool is commited (if there's no user-
overriden ratio), rather than keeping the aspect ratio of the last
rectangle (i.e. "Current"); in 2.8 this only happend when halting.
The current behavior seems more consistent anyway.


Former-commit-id: 547d3149d53d2457d9733d2a1a978a94fd34dd73