data: delete 3 brushes of very questionable origin.
It was reported that several brushes added in 1998 may have problematic
copyright with unclear/unknown licensing. And basically nobody knows
anymore where these actually come from, with which authorship or origin,
at least not for sure. It is even possible some come from commercial
software. So let's at least get rid of the ones where the origin is the
most doubtful.
Thanks to Americo for discovering these issues.


Former-commit-id: 26cb87aa2dfaaa63c71cc4e5dfe40c75f8005353